Remove old `BackendClient`
